* Added access control to Poll Creator pages

+++ web/api/creator/creator-delanswer.jsp
@@ -3,13 +3,18 @@
 <jsp:useBean id="db" scope="session" class="db.database" /> 
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%
-String[] inputData = {request.getParameter("answerID")};
-String[] inputTypes = {"int"};
+if(db.accessCheck("creator") == 1) {
+    out.print("{\"access\":\"OK\", ");
+    String[] inputData = {request.getParameter("answerID")};
+    String[] inputTypes = {"int"};
 
-db.doPreparedExecute("DELETE FROM Answers WHERE answerID=?", inputData, inputTypes);
-db.doPreparedExecute("DELETE FROM Rankings WHERE answerID=?", inputData, inputTypes);
+    db.doPreparedExecute("DELETE FROM Answers WHERE answerID=?", inputData, inputTypes);
+    db.doPreparedExecute("DELETE FROM Rankings WHERE answerID=?", inputData, inputTypes);
 
-out.print("{\"status\": \"OK\"}");
+    out.print("\"status\": \"OK\"}");
+       } else {
+            out.print("{\"access\":\"bad\"}");
+       }
 %>
